Rename and move Kind into models
There is a problem with the linter once we move sqlx_ext into this repository. Hence, this needs to be models.
